Updated the AI interfaces to use a Bridge pattern such that the IGame, IEngine, IPlayer, ICard and ITreasure classes provide a handle to the Game, Engine, Player, Card and Treasure classes which are implemented inside Domlib thus hiding the implementation but exposing the interfaces.
The AI, IAI classes work in a similar way except that the implementation of the AI class is hidden from Domlib because the AI developer writes the AI class. Most of the work of the IAI class is to translate the function calls and data structures to and from the external interface to the internal interface (i.e. ICard <-> Card, or ICardList <-> CardList).